+++ library/include/ProceduralShapeGenerators.h
@@ -88,8 +88,8 @@ public:
 	inline const ControlPoint& safeGetPoint(unsigned int i) const
 	{
 		if (mClosed)
-			return mPoints[Utils::modulo(i,mPoints.size())];
-		return mPoints[Utils::cap(i,0,mPoints.size()-1)];
+			return mPoints[Utils::modulo(i,(int)mPoints.size())];
+		return mPoints[Utils::cap(i,0,(int)mPoints.size()-1)];
 	}
 
 	/// Gets the number of control points
@@ -132,8 +132,8 @@ public:
 	inline const Ogre::Vector2& safeGetPoint(unsigned int i) const
 	{
 		if (mClosed)
-			return mPoints[Utils::modulo(i,mPoints.size())];
-		return mPoints[Utils::cap(i,0,mPoints.size()-1)];
+			return mPoints[Utils::modulo(i,(int)mPoints.size())];
+		return mPoints[Utils::cap(i,0,(int)mPoints.size()-1)];
 	}
 
 	/// Gets the number of control points
@@ -180,8 +180,8 @@ public:
 	inline const ControlPoint& safeGetPoint(unsigned int i) const
 	{
 		if (mClosed)
-			return mPoints[Utils::modulo(i,mPoints.size())];
-		return mPoints[Utils::cap(i,0,mPoints.size()-1)];
+			return mPoints[Utils::modulo(i,(int)mPoints.size())];
+		return mPoints[Utils::cap(i,0,(int)mPoints.size()-1)];
 	}
 
 	/**
@@ -481,8 +481,8 @@ public:
 	inline const Ogre::Vector2& safeGetPoint(unsigned int i) const
 	{
 		if (mClosed)
-			return mPoints[Utils::modulo(i,mPoints.size())];
-		return mPoints[Utils::cap(i,0,mPoints.size()-1)];
+			return mPoints[Utils::modulo(i,(int)mPoints.size())];
+		return mPoints[Utils::cap(i,0,(int)mPoints.size()-1)];
 	}
 
 	/// Gets the number of control points
@@ -540,8 +540,8 @@ public:
 	inline const Ogre::Vector2& safeGetPoint(unsigned int i) const
 	{
 		if (mClosed)
-			return mPoints[Utils::modulo(i,mPoints.size())];
-		return mPoints[Utils::cap(i,0,mPoints.size()-1)];
+			return mPoints[Utils::modulo(i,(int)mPoints.size())];
+		return mPoints[Utils::cap(i,0,(int)mPoints.size()-1)];
 	}
 
 	/// Gets the number of control points
